Initial commit.
This is a BIG BOI. it includes a config load system. The config load
system embeds default.hcl and then loads in configs at runtime. The
runtime configs have higher precedence.

It also includes vendoring so we dont have to worry about dependencie
version and becoming out of date.

it also includes earthly so we can have reproducible builds.