Change: manual init to DI
This is a BIGGGGGGG change. It adds DI to this code base and gets rid
manual init system.

This is a good change because it makes it easy to manage the
initialization of systems.

Aight now for the wild world of DI and that hacks that I had to do to
get it working. `tools/tools.go` is a placeholder file so code
generation actually works, googles wire lib gets in the way so this hack
is needed.

There are hacks in `Earthfile` like deleting `wire.go` so it doesnt
conflict.

Also some vendoring and stuff