Add: structured log and recovery middlewares
I added a structured logging middleware. It prints error when 5xx status
code otherwise prints as debug flags. This is important as it uses the
same logger as the rest of our program. It also lets us print more
specific information as we control the middleware.

The recovery middleware is to recover from any panics. This is important
cause if the code panics, we dont want it to straight up nuke the server.